+++ app/resources/css/editor.css
@@ -35,9 +35,6 @@ body{
 .editor-tools ul li:hover > i{
     transform: scale(1.2);
 }
-.editor-tools ul li i{
-    color: #282828;
-}
 .editor-tools ul div{
     display: flex;
     align-items: center;
@@ -73,6 +70,69 @@ body{
     border: 0;
     resize: none;
     outline: none;
-    box-sizing: border-box;
-    padding: 14px;
 }
+
+/* Spesh Switch */
+/* The switch - the box around the slider */
+.switch {
+    position: absolute;
+    right: 5%;
+    display: inline-block;
+    width: 60px;
+    height: 34px;
+  }
+  
+  /* Hide default HTML checkbox */
+  .switch input {
+    opacity: 0;
+    width: 0;
+    height: 0;
+  }
+  
+  /* The slider */
+  .slider {
+    position: absolute;
+    cursor: pointer;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background-color: rgb(255, 218, 116);
+    -webkit-transition: .4s;
+    transition: .4s;
+  }
+  
+  .slider:before {
+    position: absolute;
+    content: "";
+    height: 26px;
+    width: 26px;
+    left: 4px;
+    bottom: 4px;
+    background-color: white;
+    -webkit-transition: .4s;
+    transition: .4s;
+  }
+  
+  input:checked + .slider {
+    background-color: #414196;
+  }
+  
+  input:focus + .slider {
+    box-shadow: 0 0 1px #414196;
+  }
+  
+  input:checked + .slider:before {
+    -webkit-transform: translateX(26px);
+    -ms-transform: translateX(26px);
+    transform: translateX(26px);
+  }
+  
+  /* Rounded sliders */
+  .slider.round {
+    border-radius: 34px;
+  }
+  
+  .slider.round:before {
+    border-radius: 50%;
+  }
