+++ app/views/editor.html
@@ -63,6 +63,10 @@
                     <i class="material-icons">add</i>
                 </li>
             </div>  
+            <label class="switch">
+                <input type="checkbox">
+                <span class="slider round"></span>
+            </label>
         </ul>
     </nav>
 
@@ -77,7 +81,6 @@
     </section>
 
     <!-- Scripts -->
-    <script src="../resources/js/header.js"></script>
     <script src="../resources/js/editor.js"></script>
 </body>
 </html>
